Split scraping functions to scraper.py
